Add models

+++ db/migrate/20221206065418_create_characters_list.rb
@@ -6,15 +6,15 @@ class CreateCharactersList < ActiveRecord::Migration[7.0]
       t.string :background, null: false
       t.string :race, null: false
       t.string :alignment, null: false
-      t.decimal :exp, null: false
+      t.integer :exp, null: false
       t.boolean :inspiration, null: false
-      t.decimal :armor_class, null: false
-      t.decimal :initiative, null: false
-      t.decimal :speed, null: false
-      t.decimal :max_hp, null: false
-      t.decimal :current_hp, null: false
-      t.decimal :temporary_hp, null: false
-      t.decimal :money, null: false
+      t.integer :armor_class, null: false
+      t.integer :initiative, null: false
+      t.integer :speed, null: false
+      t.integer :max_hp, null: false
+      t.integer :current_hp, null: false
+      t.integer :temporary_hp, null: false
+      t.integer :money, null: false
 
       t.timestamps
     end
