+++ db/migrate/20221206112740_create_weapon.rb
@@ -2,8 +2,8 @@ class CreateWeapon < ActiveRecord::Migration[7.0]
   def change
     create_table :weapons do |t|
       t.string :user_view_damage, null: false
-      t.decimal :min_damage, null: false
-      t.decimal :max_damage, null: false
+      t.integer :min_damage, null: false
+      t.integer :max_damage, null: false
       t.string :type_damage, null: false
 
       t.timestamps
