+++ db/migrate/20221206112812_create_armor.rb
@@ -2,7 +2,7 @@ class CreateArmor < ActiveRecord::Migration[7.0]
   def change
     create_table :armors do |t|
       t.string :class, null: false
-      t.decimal :defend, null: false
+      t.integer :defend, null: false
 
       t.timestamps
     end
