+++ db/migrate/20221206173646_create_character_characteristics.rb
@@ -3,13 +3,13 @@ class CreateCharacterCharacteristics < ActiveRecord::Migration[7.0]
     create_table :character_characteristics do |t|
       t.references :character , null: false, foreign_key:{to_table: :characters_lists}
 
-      t.decimal :proficiency_bonus, null: false
-      t.decimal :strength, null: false
-      t.decimal :dexterity, null: false
-      t.decimal :constitution, null: false
-      t.decimal :intelligence, null: false
-      t.decimal :wisdom, null: false
-      t.decimal :charisma, null: false
+      t.integer :proficiency_bonus, null: false
+      t.integer :strength, null: false
+      t.integer :dexterity, null: false
+      t.integer :constitution, null: false
+      t.integer :intelligence, null: false
+      t.integer :wisdom, null: false
+      t.integer :charisma, null: false
 
       t.timestamps
     end
