+++ db/migrate/20221206174737_create_character_saving_throws.rb
@@ -3,12 +3,12 @@ class CreateCharacterSavingThrows < ActiveRecord::Migration[7.0]
     create_table :character_saving_throws do |t|
       t.references :character , null: false, foreign_key:{to_table: :characters_lists}
 
-      t.decimal :strength, null: false
-      t.decimal :dexterity, null: false
-      t.decimal :constitution, null: false
-      t.decimal :intelligence, null: false
-      t.decimal :wisdom, null: false
-      t.decimal :charisma, null: false
+      t.integer :strength, null: false
+      t.integer :dexterity, null: false
+      t.integer :constitution, null: false
+      t.integer :intelligence, null: false
+      t.integer :wisdom, null: false
+      t.integer :charisma, null: false
       t.timestamps
     end
   end
