+++ db/migrate/20221206192153_create_room_monsters.rb
@@ -4,7 +4,7 @@ class CreateRoomMonsters < ActiveRecord::Migration[7.0]
 
       t.references :room , null: false, foreign_key:{to_table: :rooms}
       t.references :world_monster , null: false, foreign_key:{to_table: :world_monsters}
-      t.decimal :current_hp, null: false
+      t.integer :current_hp, null: false
 
       t.timestamps
     end
